+++ management_api_app/db/errors.py
@@ -2,5 +2,9 @@ class EntityDoesNotExist(Exception):
     """Raised when entity was not found in database."""
 
 
+class EntityVersionExist(Exception):
+    """Raised when entity was not found in database."""
+
+
 class UnableToAccessDatabase(Exception):
     """Raised when we can't access the database"""
